// src/suite-vendor/a-vona/modules/a-web/src/types/request.ts
export const SymbolRequestMappingHandler = Symbol('SymbolRequestMappingHandler');

export type TypeRequestMethod = 'get' | 'post' | 'put' | 'delete' | 'patch' | 'options' | 'head';

export enum RequestMethod {
  GET = 'get',
  POST = 'post',
  PUT = 'put',
  DELETE = 'delete',
  PATCH = 'patch',
  OPTIONS = 'options',
  HEAD = 'head',
}